Enabled Collision Detection
I'm technically in charge of implementing spring physics. However, that functionality depends on collision detection, which I implemented below:

(1) Control of ball should be with respect with specific directions. Therefore, "x_ctrl" and "y_ctrl" are replaced with "left_ctrl", ... , "down_ctrl".

(2) All object coordiates (along with its length and width) are stored in a JSON called "object_coords". This enables us to easily iterate through each object's coordinates, as specified in (3)...

(3) At each "tick" in time, we iterate through each object, comparing the coordinates of the ball to that of the object. If the ball is touching the object than it cannot move in a direction that would've allowed it to move through the object.