v5
I) The current collision detection algorithm is specific to rectangular prisms. For other shapes (such as say, a pyramid), separate algorithms must be implemented.

II) Objects MUST be transformed in the specific order below in order for collision detection to work properly. Transforming in a different order affects the object's final coordinates. The collision detection algorithm assumes that the transformations are in this particular order:

1) Translation
2) Rotation
3) Scale

III) Collision detection now works for objects that are rotated AND scaled.

IV) Collision detection does NOT deflect objects realistically for now - it DOES however, deflect the object nevertheless. Implementing this algorithm is complex and requires knowing the following:

1) Which face the ball hit
2) What angle the object is rotated at
3) What direction the ball hit the object at